+++ packages/core/src/node/build.ts
@@ -62,6 +62,7 @@ export async function bundle(
             distPath: {
               root: `${outputDir}/ssr`,
             },
+            disableMinimize: true,
           },
           tools: {
             rspack(options) {
@@ -177,7 +178,9 @@ export async function renderPages(
                   helmet?.link?.toString(),
                   helmet?.style?.toString(),
                   helmet?.script?.toString(),
-                  CHECK_DARK_LIGHT_SCRIPT,
+                  config.themeConfig?.darkMode !== false
+                    ? CHECK_DARK_LIGHT_SCRIPT
+                    : '',
                 ])
                 .join(''),
             );