+++ packages/core/src/node/initRsbuild.ts
@@ -96,14 +96,15 @@ async function createInternalBuildConfig(
       setupMiddlewares: [
         middlewares => {
           if (isPublicDirExist) {
-            middlewares.push(sirv(publicDir));
+            middlewares.unshift(sirv(publicDir));
           }
 
-          middlewares.push(serveSearchIndexMiddleware(config));
+          middlewares.unshift(serveSearchIndexMiddleware(config));
         },
       ],
     },
     html: {
+      title: config?.title,
       favicon: normalizeIcon(config?.icon),
       template: path.join(PACKAGE_ROOT, 'index.html'),
     },