+++ packages/plugin-preview/src/index.ts
@@ -29,12 +29,6 @@ export type Options = {
    */
   iframePosition?: 'fixed' | 'follow';
   /**
-   * whether the demo can open in codesandbox
-   * @default false
-   * @experimental
-   */
-  enableCodesandbox?: boolean;
-  /**
    * determine how to handle a internal code block without meta
    * @default 'preview'
    */
@@ -67,7 +61,6 @@ export const demoMeta: Record<
 export function pluginPreview(options?: Options): RspressPlugin {
   const isMobile = options?.isMobile ?? false;
   const iframePosition = options?.iframePosition ?? 'follow';
-  const enableCodesandbox = options?.enableCodesandbox ?? false;
   const defaultRenderMode = options?.defaultRenderMode ?? 'preview';
 
   const demoRuntimeModule = new RspackVirtualModulePlugin({});
@@ -287,7 +280,7 @@ import Demo from ${JSON.stringify(demoComponentPath)}
       remarkPlugins: [
         [
           remarkCodeToDemo,
-          { isMobile, getRouteMeta, iframePosition, enableCodesandbox },
+          { isMobile, getRouteMeta, iframePosition, defaultRenderMode },
         ],
       ],
       globalComponents: [