+++ packages/plugin-preview/static/global-components/Container.tsx
@@ -1,22 +1,18 @@
 import { useCallback, useState } from 'react';
 import { withBase, useLang, NoSSR } from '@rspress/core/runtime';
-import { getParameters } from 'codesandbox/lib/api/define';
 import MobileOperation from './common/mobile-operation';
 import IconCode from './icons/Code';
-import IconCodesandbox from './icons/Codesandbox';
 
 type ContainerProps = {
   children: React.ReactNode[];
   isMobile: 'true' | 'false';
-  enableCodesandbox: 'true' | 'false';
   url: string;
   content: string;
   packageName: string;
 };
 
 const Container: React.FC<ContainerProps> = props => {
-  const { children, isMobile, url, content, packageName, enableCodesandbox } =
-    props;
+  const { children, isMobile, url } = props;
   const [showCode, setShowCode] = useState(false);
   const lang = useLang();
 
@@ -39,54 +35,6 @@ const Container: React.FC<ContainerProps> = props => {
     setIframeKey(Math.random());
   }, []);
 
-  const gotoCodeSandBox = () => {
-    const sandBoxConfig = {
-      files: {
-        'package.json': {
-          isBinary: false,
-          content: JSON.stringify({
-            dependencies: {
-              react: '18',
-              'react-dom': '18',
-              [packageName]: 'latest',
-            },
-          }),
-        },
-        [`demo.tsx`]: {
-          isBinary: false,
-          content,
-        },
-        [`index.tsx`]: {
-          isBinary: false,
-          content: [
-            `import React from 'react'`,
-            `import ReactDOM from 'react-dom'`,
-            `import Demo from './demo'`,
-            `ReactDOM.render(<Demo />, document.getElementById('root'))`,
-          ].join('\n'),
-        },
-      },
-    };
-
-    // to specific demo file
-    const query = `file=/demo.tsx`;
-    const form = document.createElement('form');
-    form.action = `https://codesandbox.io/api/v1/sandboxes/define?query=${encodeURIComponent(
-      query,
-    )}`;
-    form.target = '_blank';
-    form.method = 'POST';
-    form.style.display = 'none';
-    const field = document.createElement('input');
-    field.name = 'parameters';
-    field.type = 'hidden';
-    field.setAttribute('value', getParameters(sandBoxConfig));
-    form.appendChild(field);
-    document.body.appendChild(form);
-    form.submit();
-    document.body.removeChild(form);
-  };
-
   return (
     <NoSSR>
       <div className="rspress-preview">
@@ -95,13 +43,7 @@ const Container: React.FC<ContainerProps> = props => {
             <div className="rspress-preview-code">{children?.[0]}</div>
             <div className="rspress-preview-device">
               <iframe src={getPageUrl()} key={iframeKey}></iframe>
-              <MobileOperation
-                url={url}
-                refresh={refresh}
-                gotoCodeSandBox={
-                  enableCodesandbox === 'true' ? gotoCodeSandBox : undefined
-                }
-              />
+              <MobileOperation url={url} refresh={refresh} />
             </div>
           </div>
         ) : (
@@ -116,18 +58,6 @@ const Container: React.FC<ContainerProps> = props => {
                 {children?.[1]}
               </div>
               <div className="rspress-preview-operations web">
-                {enableCodesandbox === 'true' && (
-                  <button
-                    onClick={gotoCodeSandBox}
-                    aria-label={
-                      lang === 'zh'
-                        ? '在 Codesandbox 打开'
-                        : 'Open in Codesandbox'
-                    }
-                  >
-                    <IconCodesandbox />
-                  </button>
-                )}
                 <button
                   onClick={toggleCode}
                   aria-label={lang === 'zh' ? '收起代码' : 'Collapse Code'}