+++ packages/theme-default/src/index.ts
@@ -13,6 +13,8 @@ export { Link } from './components/Link';
 export { HomeFooter } from './components/HomeFooter';
 export { Toc } from './components/Toc';
 export { PackageManagerTabs } from './components/PackageManagerTabs';
+export { LastUpdated } from './components/LastUpdated';
+export { PrevNextPage } from './components/PrevNextPage';
 
 export default {
   Layout,