+++ packages/core/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @rspress/core
 
+## 1.8.2
+
+### Patch Changes
+
+- @rspress/plugin-auto-nav-sidebar@1.8.2
+- @rspress/plugin-container-syntax@1.8.2
+- @rspress/plugin-last-updated@1.8.2
+- @rspress/plugin-medium-zoom@1.8.2
+- @rspress/runtime@1.8.2
+- @rspress/shared@1.8.2
+- @rspress/theme-default@1.8.2
+
+## 1.8.1
+
+### Patch Changes
+
+- d6f5294: feat: use config.title as html title.
+- c8e655e: perf(deps): remove unused jsdom dependency
+- 21b2845: fix: disable minify in ssr build
+- 0a4b727: feat: support url query to disable global ui components
+- f7a5a09: fix(core): stringify global compoennts path
+- c038048: chore(deps): bump Rsbuild from v0.1.0 to v0.1.9
+- Updated dependencies [126f0aa]
+- Updated dependencies [a96e651]
+- Updated dependencies [c8e655e]
+- Updated dependencies [df6e061]
+- Updated dependencies [0a4b727]
+- Updated dependencies [c038048]
+  - @rspress/runtime@1.8.1
+  - @rspress/shared@1.8.1
+  - @rspress/theme-default@1.8.1
+  - @rspress/plugin-medium-zoom@1.8.1
+  - @rspress/plugin-auto-nav-sidebar@1.8.1
+  - @rspress/plugin-container-syntax@1.8.1
+  - @rspress/plugin-last-updated@1.8.1
+
+## 1.8.0
+
+### Patch Changes
+
+- e4329b6: fix(core): support custom id when disable mdxRs and escape heading id in processor
+- bfc0e80: fix: devServer use searchIndexMiddleware first.
+- 7841e4c: fix: disable dark mode script when darkMode is false in config
+- Updated dependencies [482faf5]
+- Updated dependencies [b435c24]
+- Updated dependencies [3756fab]
+  - @rspress/shared@1.8.0
+  - @rspress/theme-default@1.8.0
+  - @rspress/plugin-auto-nav-sidebar@1.8.0
+  - @rspress/plugin-container-syntax@1.8.0
+  - @rspress/plugin-last-updated@1.8.0
+  - @rspress/plugin-medium-zoom@1.8.0
+  - @rspress/runtime@1.8.0
+
 ## 1.7.5
 
 ### Patch Changes