+++ packages/plugin-container-syntax/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rspress/remark-container
 
+## 1.8.2
+
+### Patch Changes
+
+- @rspress/shared@1.8.2
+
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies [126f0aa]
+- Updated dependencies [0a4b727]
+- Updated dependencies [c038048]
+  - @rspress/shared@1.8.1
+
+## 1.8.0
+
+### Patch Changes
+
+- Updated dependencies [482faf5]
+  - @rspress/shared@1.8.0
+
 ## 1.7.5
 
 ### Patch Changes