+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @rspress/runtime
 
+## 1.8.2
+
+### Patch Changes
+
+- @rspress/shared@1.8.2
+
+## 1.8.1
+
+### Patch Changes
+
+- 126f0aa: fix: inline image data url start with slash prefix.
+- Updated dependencies [126f0aa]
+- Updated dependencies [0a4b727]
+- Updated dependencies [c038048]
+  - @rspress/shared@1.8.1
+
+## 1.8.0
+
+### Patch Changes
+
+- Updated dependencies [482faf5]
+  - @rspress/shared@1.8.0
+
 ## 1.7.5
 
 ### Patch Changes